+++ adapters/repos/db/vector/hnsw/commit_logger_noop.go
@@ -33,15 +33,15 @@ func (n *NoopCommitLogger) Flush() error {
 	return nil
 }
 
-func (n *NoopCommitLogger) SetEntryPointWithMaxLayer(id uint64, level int) error {
+func (n *NoopCommitLogger) SetEntryPointWithMaxLayer(id uint64, level int8) error {
 	return nil
 }
 
-func (n *NoopCommitLogger) AddLinkAtLevel(nodeid uint64, level int, target uint64) error {
+func (n *NoopCommitLogger) AddLinkAtLevel(nodeid uint64, level int8, target uint64) error {
 	return nil
 }
 
-func (n *NoopCommitLogger) ReplaceLinksAtLevel(nodeid uint64, level int, targets []uint64) error {
+func (n *NoopCommitLogger) ReplaceLinksAtLevel(nodeid uint64, level int8, targets []uint64) error {
 	return nil
 }
